Create photo gallery container
